pyproject.toml
[project]
name = "pymap"
version = "0.0.1"
description = "PyLens helps you find PyPI packages through natural language prompts with the help of Large Language Models (LLMs)."
authors = [{ name = "Florian Maas", email = "ffpgmaas@gmail.com" }]
requires-python = ">=3.9,<4.0"
readme = "README.md"
dependencies = [
    "beautifulsoup4>=4.12.3,<5",
    "polars>=0.20.31,<0.21",
    "sentence-transformers>=3.0.1,<4",
    "lxml>=5.2.2,<6",
    "python-dotenv>=1.0.1,<2",
    "tqdm>=4.66.4,<5",
    "fastapi>=0.111.0,<0.112",
    "pydantic>=2.7.4,<3",
    "pydantic-settings>=2.0.0,<3",
    "uvicorn>=0.30.1,<0.31",
    "gdown>=5.2.0,<6",
    "azure-storage-blob>=12.20.0,<13",
    "slowapi>=0.1.9,<0.2",
    "starlette>=0.37.2,<0.38",
    "numpy>=2.0.0,<3",
    "scikit-learn>=1.5.0,<2",
    "readme-renderer[md]>=44.0",
    "plotly>=6.5.0",
    "pyarrow>=21.0.0",
    "pandas>=2.3.3",
    "umap-learn>=0.5.9.post2",
    "hdbscan>=0.8.40",
]

[project.urls]
Repository = "https://github.com/fpgmaas/pymap"
Documentation = "https://fpgmaas.github.io/pymap/"

[dependency-groups]
dev = [
    "pytest>=7.2.0,<8",
    "pytest-cov>=4.0.0,<5",
    "pytest-mock>=3.14.0,<4",
    "deptry>=0.12.0,<0.13",
    "pre-commit>=3.4.0,<4",
    "tox>=4.11.1,<5",
    "marimo>=0.17.6",
]

[tool.hatch.build.targets.sdist]
include = ["pymap"]

[tool.hatch.build.targets.wheel]
include = ["pymap"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py37"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    "TRY003",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["pymap"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.deptry]
extend_exclude = [
    "frontend"
]

[tool.deptry.per_rule_ignores]
DEP002 = ["lxml", "uvicorn"]
